requirements.txt
requests
polars
python-dotenv
